--- nvim/lua/exp/init.lua
@@ -1,20 +1,0 @@
-require('exp.set')
-
-local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
-if not vim.loop.fs_stat(lazypath) then
-	vim.fn.system({
-		"git",
-		"clone",
-		"--filter=blob:none",
-		"https://github.com/folke/lazy.nvim.git",
-		"--branch=stable",
-		lazypath,
-	})
-end
-vim.opt.rtp:prepend(lazypath)
-
-require('lazy').setup('exp.plugins')
-require('exp.lsp')
-require('exp.remap')
-require('exp.lazygit')
-require('exp.lazytest')
