--- nvim/lua/exp/plugins/fidget.lua
@@ -1,11 +1,0 @@
-return {
-    'j-hui/fidget.nvim',
-    opts = {
-        notification = {
-            window = {
-                winblend = 0,
-                border = 'rounded',
-            },
-        },
-    },
-}
