--- nvim/lua/exp/plugins/gitsigns.lua
@@ -1,7 +1,0 @@
-return {
-    'lewis6991/gitsigns.nvim',
-    config = function()
-        require('gitsigns').setup()
-        vim.cmd('Gitsigns toggle_current_line_blame')
-    end
-}
