--- nvim/lua/exp/plugins/surround.lua
@@ -1,7 +1,0 @@
-return {
-    'kylechui/nvim-surround',
-    version = '*',
-    config = function()
-        require('nvim-surround').setup()
-    end
-}
