--- nvim/lua/exp/plugins/toggleterm.lua
@@ -1,5 +1,0 @@
-return {
-    'akinsho/toggleterm.nvim',
-    version = '*',
-    config = true,
-}
