+++ nvim/lua/kampanosg/init.lua
@@ -1,20 +1,20 @@
-require('kampanosg.set')
+require('exp.set')
 
 local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
 if not vim.loop.fs_stat(lazypath) then
-	vim.fn.system({
-		"git",
-		"clone",
-		"--filter=blob:none",
-		"https://github.com/folke/lazy.nvim.git",
-		"--branch=stable",
-		lazypath,
-	})
+    vim.fn.system({
+        "git",
+        "clone",
+        "--filter=blob:none",
+        "https://github.com/folke/lazy.nvim.git",
+        "--branch=stable",
+        lazypath,
+    })
 end
 vim.opt.rtp:prepend(lazypath)
 
-require('lazy').setup('kampanosg.plugins')
-require('kampanosg.lsp')
-require('kampanosg.remap')
-require('kampanosg.lazygit')
-require('kampanosg.lazytest')
+require('lazy').setup('exp.plugins')
+require('exp.lsp')
+require('exp.remap')
+require('exp.lazygit')
+require('exp.lazytest')
