+++ nvim/lua/kampanosg/lsp.lua
@@ -1,56 +1,81 @@
-local lsp = require("lsp-zero")
+local lsp = require('lspconfig')
 
-lsp.preset("recommended")
-
-lsp.ensure_installed({
-    'tsserver',
-    'eslint',
-    'rust_analyzer',
+local language_servers = {
     'gopls',
-})
+    'rust_analyzer',
+    'lua_ls',
+    'bashls',
+    'dockerls',
+    'tsserver',
+    'yamlls',
+    'terraform_lsp',
+}
 
-local cmp = require('cmp')
-local cmp_select = { behavior = cmp.SelectBehavior.Select }
-local cmp_mappings = lsp.defaults.cmp_mappings({
-    ['<C-p>'] = cmp.mapping.select_prev_item(cmp_select),
-    ['<C-n>'] = cmp.mapping.select_next_item(cmp_select),
-    ['<CR>'] = cmp.mapping.confirm({ select = true }),
-    ['<C-Space>'] = cmp.mapping.complete(),
-})
+local capabilities = vim.lsp.protocol.make_client_capabilities()
+capabilities.textDocument.foldingRange = {
+    dynamicRegistration = false,
+    lineFoldingOnly = true
+}
 
-lsp.set_preferences({
-    sign_icons = {}
-})
+for _, ls in ipairs(language_servers) do
+    lsp[ls].setup({
+        on_attach = on_attach,
+        capabilities = capabilities
+    })
+end
 
-lsp.setup_nvim_cmp({
-    mapping = cmp_mappings
-})
+lsp.tsserver.setup {
+    init_options = {
+        plugins = {
+            {
+                name = '@vue/typescript-plugin',
+                location = '/usr/local/lib/node_modules/@vue/typescript-plugin',
+                languages = { 'javascript', 'typescript', 'vue' },
+            },
+        },
+    },
+    filetypes = {
+        'javascript',
+        'typescript',
+        'vue',
+    },
+}
 
-lsp.on_attach(function(client, bufnr)
-    local opts = { buffer = bufnr, remap = false }
-
-    vim.keymap.set("n", "<leader>gR", function() vim.lsp.buf.rename() end, opts)
-    vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
-    vim.keymap.set('n', '<C-k>', vim.lsp.buf.signature_help, opts)
-    vim.keymap.set("n", "<F2>", function() vim.diagnostic.goto_next() end, opts)
-    vim.keymap.set("n", "<F3>", function() vim.diagnostic.goto_prev() end, opts)
-end)
-
-lsp.setup()
-
-
--- format Go on save
-local format_sync_grp = vim.api.nvim_create_augroup("GoImport", {})
-vim.api.nvim_create_autocmd("BufWritePre", {
-  pattern = "*.go",
-  callback = function()
-   require('go.format').goimport()
-  end,
-  group = format_sync_grp,
-})
-
+-- format on save
 vim.api.nvim_create_autocmd('BufWritePre', {
-  pattern = { '*.tsx', '*.ts', '*.jsx', '*.js' },
-  command = 'silent! EslintFixAll',
-  group = vim.api.nvim_create_augroup('MyAutocmdsJavaScripFormatting', {}),
+    buffer = buffer,
+    callback = function()
+        vim.lsp.buf.format { async = false }
+    end
+})
+
+vim.api.nvim_create_autocmd('LspAttach', {
+    group = vim.api.nvim_create_augroup('gk-lsp-attach', { clear = true }),
+    callback = function(event)
+        local client = vim.lsp.get_client_by_id(event.data.client_id)
+        if client == nil then
+            print("client is nil")
+            return
+        end
+
+        -- highlight same symbols under cursor
+        if client and client.server_capabilities.documentHighlightProvider then
+            vim.api.nvim_create_autocmd({ 'CursorHold', 'CursorHoldI' }, {
+                buffer = event.buf,
+                callback = vim.lsp.buf.document_highlight,
+            })
+
+            vim.api.nvim_create_autocmd({ 'CursorMoved', 'CursorMovedI' }, {
+                buffer = event.buf,
+                callback = vim.lsp.buf.clear_references,
+            })
+        end
+
+        -- keymaps
+        local opts = { noremap = true, silent = true }
+        vim.keymap.set('n', '<leader>gR', function() vim.lsp.buf.rename() end, opts)
+        vim.keymap.set('n', '<F2>', function() vim.diagnostic.goto_next() end, opts)
+        vim.keymap.set('n', '<F3>', function() vim.diagnostic.goto_prev() end, opts)
+        vim.keymap.set('n', '<F4>', function() vim.diagnostic.open_float() end, opts)
+    end
 })
