--- nvim/lua/kampanosg/plugins/comment.lua
@@ -1,5 +1,0 @@
-return {
-    'numToStr/Comment.nvim',
-    opts = {},
-    lazy = false,
-}
