+++ nvim/lua/kampanosg/plugins/fidget.lua
@@ -4,6 +4,7 @@ return {
         notification = {
             window = {
                 winblend = 0,
+                border = 'rounded',
             },
         },
     },
