+++ nvim/lua/kampanosg/plugins/surround.lua
@@ -1,7 +1,6 @@
 return {
     'kylechui/nvim-surround',
     version = '*',
-    -- event = 'VeryLazy',
     config = function()
         require('nvim-surround').setup()
     end
